orchestrator/requirements.txt
httpx>=0.27.0
pydantic>=2.7.0

# Azure AI Foundry SDK
azure-ai-projects>=2.0.0
azure-identity>=1.17.0

# dev / test
pytest>=8.0.0
pytest-anyio>=0.0.0
anyio>=4.0.0
